New constructor (DList) and add three functions (getNext, getPnext, getKey)
